text_rpg_rust::models::entities::character
Struct Character

pub struct Character {
    pub(crate) vivant: Vivant,
    pub position: usize,
    pub level: i32,
    pub money: i32,
    pub quests: Vec<u32>,
}

Fields
vivant: Vivantposition: usizelevel: i32money: i32quests: Vec<u32>
Implementations

impl Character

pub fn try_move(&mut self, direction: &str, rooms: &mut [Room])

pub fn prendre_objet( &mut self, objet_nom: &str, rooms: &mut [Room], items: &[Item], quetes: &mut HashMap<u32, Quete>, dialogues: &mut Vec<Dialogue>, )

pub fn utiliser_objet( &mut self, objet_nom: &str, rooms: &mut [Room], items: &[Item], )

pub fn level_up(&mut self)

pub fn add_experience(&mut self, xp: i32)

pub fn afficher_inventaire(&self, items: &[Item])

pub fn ajouter_quete(&mut self, id: u32)

pub fn supprimer_quete(&mut self, id: u32)

pub fn id(&self) -> u32

pub fn name(&self) -> &str

pub fn health(&self) -> i32

pub fn strength(&self) -> i32

pub fn intelligence(&self) -> i32

pub fn set_health(&mut self, health: i32)

pub fn set_strength(&mut self, strength: i32)

pub fn set_intelligence(&mut self, intelligence: i32)

pub fn inventory_mut(&mut self) -> &mut Inventory

pub fn defense(&self) -> i32

pub fn is_alive(&self) -> bool

pub fn get_item_details<'a>( &self, item_id: u32, items: &'a [Item], ) -> Option<&'a Item>

pub fn get_active_quests( &self, all_quests: &HashMap<u32, Quete>, items: &Vec<Item>, enemies: &HashMap<u32, Enemy>, ) -> Vec<String>

pub fn quests(&self) -> &Vec<u32>

pub fn add_argent(&mut self, quantité: i32)

pub fn afficher_et_ajouter_loot( &mut self, loot: &[InventoryItem], items: &[Item], )

pub fn combat_interactif<T: Combattant>( &mut self, ennemi: &mut T, items: &[Item], ) -> CombatResult

Lance un combat interactif entre le joueur et un ennemi donné.

Le joueur peut choisir d’attaquer, d’utiliser un objet ou de fuir. Si le joueur gagne, il obtient du loot et de l’expérience.

Paramètres

ennemi : l’ennemi à combattre.

items : la liste globale des objets connus du jeu.

Retour

Un CombatResult indiquant si le joueur a gagné, perdu, ou fui.